Models updated

+++ Clients/AddSubscriptionTypeRequest.cs
@@ -9,6 +9,6 @@ namespace Goova.Subscriptions.Models.Clients
         public string Name { get; set; }
         public string Description { get; set; }
         public int Frequence { get; set; }
-        public double SubscriptionCost { get; set; }
+        public decimal SubscriptionCost { get; set; }
     }
 }
